server/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
mongoengine==0.29.1
PyJWT==2.8.0
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.13.1
python-dotenv==1.1.0
python_bcrypt==0.3.2
